PR fixes, removed cycle
