+++ spring-cloud-gcp-data-spanner/pom.xml
@@ -44,11 +44,6 @@
 			<artifactId>spring-tx</artifactId>
 			<scope>test</scope>
 		</dependency>
-		<dependency>
-			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-gcp-autoconfigure</artifactId>
-			<scope>test</scope>
-		</dependency>
 	</dependencies>
 	<build>
 		<plugins>
