+++ spring-cloud-gcp-samples/spring-cloud-gcp-data-spanner-sample/src/main/java/com/example/SpannerApplication.java
@@ -18,11 +18,13 @@ package com.example;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.gcp.data.spanner.repository.config.EnableSpannerRepositories;
 
 /**
  * @author Chengyuan Zhao
  */
 @SpringBootApplication
+@EnableSpannerRepositories(namedQueriesLocation = "classpath:spanner-named-queries.properties")
 public class SpannerApplication {
 
 	public static void main(String[] args) {
